fix: adjust naming of KMS outputs
I'm adjusting the naming of the KMS outputs so they all start with a `kms_` prefix. I also adjusted the name of the `encryption_key_arn` variable in the S3 module to `kms_master_key_arn`. This was to make it clear that it's the key used for infra level encryption to be used, and not our in-app encryption key.