feat: update the module to support custom S3 expirations
Our existing CloudFormation approach allows customers to adjust the retention periods of certain buckets. I've updated the module to allow retention periods to be passed in to help with v2 -> v3 migrations.

I also noticed that the default value for the `rds_backup_retention_period` didn't match the current v2 approach, so I've increased it to match.